DNS Answers: Add columns for AS number and organization name
The measurement JSON may include additional information about IP
addresses returned in DNS answers.  In particular, the "asn" field
contains the AS (Autonomous System) number advertising the IP address,
and "as_org_name" the organizational name for that AS.

This change adds these fields to the answer table as two additional
columns.